Fix concurrency issue in mutlithreaded environment (2.rc.03)
Problem:
If two threads called base64_encode concurrently (one with url=true and
the other with url=false, the values of base64_chars[62] and
base64_chars[63] are undefined. For example, it's possible to get
base64_chars[62]='-' and base64_chars[63]='/', which is not a valid
encoding.

This commit fixes this issue.